backend/requirements.txt
django==3.2.14
django-environ==0.9.0
django-model-utils==4.2.0
django-allauth==0.51.0
django-crispy-forms==1.14.0
crispy-bootstrap5==0.6
django-redis==5.2.0
djangorestframework==3.13.1
django-cors-headers==3.13.0
drf-spectacular==0.22.1
django-anymail==8.6
gunicorn==20.1.0
psycopg2==2.9.3
sentry-sdk==1.6.0
django-modeltranslation==0.18.5
dj-rest-auth==2.2.5
google-auth==2.15.0
django-simple-history==3.2.0
rules==3.3
django-filter==22.1
fhirclient==4.1.0
fhir.resources==7.1.0
django-leaflet==0.29.0
djangorestframework-gis==1.0
django-storages==1.13.2
channels==4.0.0
channels-redis==4.1.0
urllib3==1.26
pytz==2022.1
python-slugify==6.1.2
argon2-cffi==21.3.0
whitenoise==6.2.0
redis==4.5.3
hiredis==2.0.0
uvicorn[standard]==0.17.6
factory-boy==3.2.1
ipdb
pandas
numpy
googlemaps

